chore: update tailwindcss and vite versions
