+++ package.json
@@ -4,6 +4,7 @@
     "description": "DD360 Components is a tailwind based React component library",
     "license": "MIT",
     "homepage": "https://bui.dd360.mx",
+    "type": "module",
     "repository": {
         "type": "git",
         "url": "https://github.com/dd3tech/bui"
@@ -91,7 +92,8 @@
         "@types/react-dom": "17.0.0",
         "@typescript-eslint/eslint-plugin": "^5.56.0",
         "@typescript-eslint/parser": "^5.56.0",
-        "@vitejs/plugin-react": "2.0.0",
+        "@vitejs/plugin-react": "4.3.1",
+        "@vitest/coverage-v8": "^1.6.0",
         "autoprefixer": "10.4.0",
         "babel-loader": "8.2.3",
         "chromatic": "6.7.4",
@@ -120,10 +122,10 @@
         "rollup-plugin-typescript2": "0.32.1",
         "rollup-plugin-visualizer": "4.2.0",
         "semantic-release": "^19.0.5",
-        "tailwindcss": "2.2.19",
+        "tailwindcss": "3.4.4",
         "typescript": "4.3.2",
-        "vite": "3.0.4",
-        "vitest": "^0.20.3"
+        "vite": "5.3.1",
+        "vitest": "1.6.0"
     },
     "config": {
         "commitizen": {
